Monarch web-crawler example. (#1359)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1359

This web-crawler example takes a target site and crawls it to a depth of 3.  It is initialized by adding the root to a central/single queue actor, and a number of parallel crawling actors take off the queue, crawl, and add to the queue, until the depth is exhausted.  The crawlers only crawl sub-pages of the base URL (just for safety/simplicity), and stop at the target depth.

Reviewed By: colin2328

Differential Revision: D83214162

fbshipit-source-id: 65f6e753bcfc9bce6982e1e329d6abed5f369660