Turn on cargo test for the hyperactor crate (#1417)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1417

We should have Rust tests run in Github CI to ensure we don't break them.
Before we were blocked on issues with `cargo test` running tests in threads, and we
need process isolation because our tests modify things like environment variables.

Use `cargo nextest` which is a package that runs test in parallel processes instead of
threads.

For now, just run these in a GPU environment. A cpu-only run may be beneficial for either
cost savings, faster iteration cycles, or for test coverage itself. But I'll add that later.

Some tests are failing in Github but not internally. Add a feature to skip these called "fb".

Reviewed By: mariusae

Differential Revision: D83856634

fbshipit-source-id: 87c1fe1053e092fc0c4076bca5856912be3c4ea0