Lazy NIC device init + integration of NIC autodetection. (#1444)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1444

Rewrites RdmaManager internals wrt:

Before - we select a given RDMA-compliant device on (on init via config) and use that for all network traffic.

Now - RdmaManager is a general resource manager for a given process.   For a given memory virtual addr, we will select the closest RDMA device (pci-distance), and setup the relevant QPs, pd, on the fly.   So a given manager could be sending on traffic on "N" rdma devices.

To make this work, we need to add a 'device_name' field to RdmaBuffer, so its a combination of other_id + other_device; introducing a many to many connection setup between two rdma managers.

Reviewed By: casteryh

Differential Revision: D83918488

fbshipit-source-id: f8b7e5cd1abe29e08afbcc526ab02053c9fa12f5