Enable pickling HostMesh and ProcMesh from inside tokio thread in certain cases (#1473)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1473

Previously, both `HostMesh` and `ProcMesh` stored their underlying rust mesh in an asynchronous `Shared[...]` pytokio object, so any attempts to access the rust meshes from a tokio thread would need to be called from a coroutine. This is a problem for pickling, which needs to be synchronous, and would therefore have to call `Shared[...].block_on()` to get the underlying rust mesh.

This diff makes it so that when the internal `Shared[...]` task for `HostMesh` and `ProcMesh` completes, the result is stored so that it can be accessed without the use of `block_on()`. This enables pickling `HostMesh` and `ProcMesh` inside tokio threads, as long as their backing rust meshes have finished initializing -- this will always be the case inside actor endpoints, since the `HostMesh` and `ProcMesh` both need to be done initializing by the time the actor endpoint runs.
ghstack-source-id: 315174591
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D84195494

fbshipit-source-id: 08f0f6f0ece9806a93884e18da933b4bef8f82e6