Fix too many tasks with supervision (#1514)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1514

Part of: https://github.com/meta-pytorch/monarch/issues/1209

There was a performance issue with a polling-based supervision events, where the
periodic timer loop would run forever, even if the endpoint which generated it
had been deleted.

The cause was my misunderstanding of how PyShared worked, it is a way to turn a one-off
task into a continuous task.
It does this:
```lang=rs
get_tokio_runtime().spawn(async move {
  send_result(tx, task.await);
});
```
The problem was that this tokio task lived until the `.await` completed. In the case of
supervision on a healthy endpoint, that await would never complete.
For v1, this meant that the polling loop would never get dropped.
Using a OnceLock to keep it per-mesh is better in that it doesn't grow with each endpoint
call, but it would still grow with each MeshRef and slice. So we need a way to clean up these tasks
if nothing is going to await them.

This problem affects v0 supervision as well, where the number of alive tokio tasks grew without bound;
however, in v0 the individual tasks didn't use any CPU or memory resources, they just
waited on a channel. In v1 they now send out periodic "cast" messages to collect state, so
we don't want them hanging around forever.

To this end, add a `PyPythonTask::spawn_abortable` to turn a task that is safe to cancel
into a PyShared. When the PyShared is dropped, it aborts the task.
This change allows the supervision monitors to be dropped as soon as the endpoint
completes in a healthy manner.

Reviewed By: mariusae

Differential Revision: D84400989

fbshipit-source-id: a0702dbba36f9d207dfd73474d7ec080959d152d