Implement __supervise__ callback on PythonActor (#1547)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1547

Part of https://github.com/meta-pytorch/monarch/issues/1209

The goal of the v1 supervision API is to give owning PythonActors a chance to recover
from a failure on a mesh they own.
To this end, we want to invoke a `__supervise__` function on the PythonActor whenever
an actor it "owns" changes state. For now we're mostly interested in state changes to "Failed"
or "Stopped".

To do so, we reuse the continuous polling we did for the port monitoring for v0 supervision
events, and additionally send a message to an owning PythonActor.
If the owner is a root client actor (the unit actor `()`), we don't send this message for now.

We provide a `Failure` exception object with the mesh that failed. It's `__str__` will include
the rank, actorid, and as much information as we have. For now those properties are not
directly exposed to allow us flexibility in evolving the API.

The `__supervise__` return API is similar to `__exit__` for context managers:
Truthy return value suppresses original supervision error (handled), Falsey return value allows it to bubble (unhandled). Exceptions raised from the callback become their own supervision event,
whose cause is the previous supervision event.

Some limitations of this implementation:
1. Currently the "owning" actor is fetched via the context at the time of the endpoint call. This means that if a ref is passed to another actor, that actor will receive the `__supervise__`, and not the original creator. This is not the intended long term behavior, as refs should only get the "instantaneous" exceptions and not a notification.
2. Bubbling up to the next owning actor is not possible because it's not reachable from an Instance.
We need to have a way to get an ActorRef to the original owner. Solving (1) should make this solvable too.
3. Client-owned meshes don't get these events delivered, because there is no `__supervise__` to call. In the future, we may want to give users a way to provide a callback function to spawn

Reviewed By: mariusae

Differential Revision: D84020973

fbshipit-source-id: ef96e814deb8debdb6757d9ab08d65da3fff2bae