Add stop for v1::ActorMesh (#1529)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1529

Implement `.stop()` for an ActorMesh. This sends a message to the ProcMeshAgent
to stop the actors on all procs that the actors live on.

This is done with a new "Destroy" message that is the parallel of CreateOrUpdate.
It will be implemented for HostMesh next to terminate procs.
In order to cast a Destroy message, it needs a context::Actor impl which it gets from
an Instance object.

If any actor is used after stop, it'll get a NotExist error.
Requires supervision to be enabled in order to get that error correctly without a timeout.

Reviewed By: mariusae

Differential Revision: D84547919

fbshipit-source-id: 62baacc5d2ddefa7b9bf2425cec99b62df43eb27