Plumb global config through from client to host (#1521)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1521

This diff adds new global config layer `ClientOverride`. When a new proc mesh is created, the current value of the global config on the client is sent to each host mesh agent and passed to the agent's `BootstrapProcManager`. When the `BootstrapProcManager` spawns a new process, the config snapshot is set in that process's global config at the `ClientOverride` layer, which is lower priority than any other layer (to account for, say, the user explicitly setting env vars on their hosts).

I plumbed this through a separate path than `BootstrapCommand` so that it's applicable for both the proc mesh trampoline approach and the simple bootstrap approach.
ghstack-source-id: 317272326
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D84578137

fbshipit-source-id: faa698d141636c286f57a4fa82cf39e7db0b86fe