Downgrade message log level to TRACE (#1672)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1672

We found this log not helpful when debugging S576170. But it causes log spew and makes it hard to find other useful logs.

Downgrade to TRACE instead of deleting it just in case we might need it in the future debugging (at that time we can manually bump its level).

Reviewed By: shayne-fletcher

Differential Revision: D85533659

fbshipit-source-id: 1b07acd5dbee9245fc305601b8366eed96998492