mesh: HostMesh::process, test_bench (#1661)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1661

HostMesh::process constructs a host mesh by forking a process for each host in the mesh. It can be used to bootstrap a large host mesh from a single binary on a single host. The intent is to use for testing, benchmarking, tool development, etc.

test_bench is meant as a template "test bench" for purposes of testing, development, and benchmarking.
ghstack-source-id: 319505789
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D85476028

fbshipit-source-id: e0e2fa4f745ad5a0cfd8e23480dd076f7d2c40f6