Ignore mailbox errors in GetState for HostMeshAgent and ProcMeshAgent (#1691)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1691

If some actor owns some ActorMeshes, it will periodically send out a GetState message
to the HostMeshAgent and ProcMeshAgent. If that sender crashes while waiting for a reply,
it'll cause a MailboxSenderError on the agents.
We don't want those agents to stop because of such an error, as it just means nobody will
receive a reply.

GetRankStatus and GetState messages are read-only and have no side effects, so it's fine
to just warn on the MailboxSenderError, there is no invalid state left hanging around.

Reviewed By: vidhyav, xunnanxu

Differential Revision: D85720817

fbshipit-source-id: 4e2881553590b08e60ca7a264eb669fa7ff082ec